+++ R/namespace.R
@@ -1,4 +1,5 @@
 #' @import chk term extras
+#' @import lifecycle
 #' @importFrom stats aggregate
 #' @importFrom abind abind
 #' @importFrom purrr transpose