+++ R/natomic-object.R
@@ -8,8 +8,8 @@
 #' @examples
 #' mat <- matrix(c("1", "2", "3", "4"), 2L)
 #' mat
-#' is.natomic(mat)
+#' is_natomic(mat)
 #' mat <- extras::numericize(mat)
 #' mat
-#' is.natomic(mat)
+#' is_natomic(mat)
 NULL