+++ R/params.R
@@ -1,0 +1,8 @@
+#' Parameter Descriptions
+#'
+#' @param x An object.
+#' @param flesh An atomic vector
+#' @param skeleton An nlist object.
+#'
+#' @name params
+NULL